+++ src/main/java/com/infeco/keylease/entity/PropertyEntity.java
@@ -35,6 +35,9 @@ public class PropertyEntity {
     )
     private PropertyTypeEntity propertyType;
 
+    @OneToOne(mappedBy = "property")
+    private LeaseContractEntity leaseContract;
+
     public PropertyEntity() {
     }
 
@@ -84,5 +87,13 @@ public class PropertyEntity {
 
     public void setPropertyType(PropertyTypeEntity propertyType) {
         this.propertyType = propertyType;
+    }
+
+    public LeaseContractEntity getLeaseContract() {
+        return leaseContract;
+    }
+
+    public void setLeaseContract(LeaseContractEntity leaseContract) {
+        this.leaseContract = leaseContract;
     }
 }